src/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
pgvector
python-jose
passlib[bcrypt]
python-multipart
python-dotenv
langchain-huggingface
sentence-transformers
scikit-learn
together
pydantic
typing-extensions
uv

prometheus-client
fastapi-health
starlette-exporter